Move get_status_dict method to a separate utils module
