+++ traitcuration/traits/utils.py
@@ -1,0 +1,16 @@
+def get_status_dict(traits):
+    status_dict = {
+        "all": {"count": len(traits), "class": "primary"},
+        "awaiting_review": {"count": 0, "class": "warning"},
+        "awaiting_creation": {"count": 0, "class": "warning"},
+        "needs_creation": {"count": 0, "class": "warning"},
+        "awaiting_import": {"count": 0, "class": "warning"},
+        "needs_import": {"count": 0, "class": "warning"},
+        "unmapped": {"count": 0, "class": "danger"},
+        "obsolete": {"count": 0, "class": "danger"},
+        "deleted": {"count": 0, "class": "danger"},
+        "current": {"count": 0, "class": "success"},
+    }
+    for trait in traits:
+        status_dict[trait.status]["count"] += 1
+    return status_dict
